+++ spec/scopes_spec.rb
@@ -10,11 +10,26 @@ describe "scopes" do
     expect(CashRegister.unknown).to be_empty
   end
 
+  it "builds scopes for inverse elements" do
+    unknowns = CashRegister.create!(status: CashRegister::UNKNOWN)
+    actives = CashRegister.create!(status: CashRegister::ACTIVE)
+    alarms = CashRegister.create!(status: CashRegister::ALARM)
+
+    expect(CashRegister.known).to contain_exactly(actives, alarms)
+  end
+
   it "builds scopes with prefixed names for each namespaced element" do
     opened = CashRegister.new.tap(&:open!)
     closed = CashRegister.new.tap(&:close!)
 
     expect(CashRegister.drawer_position_opened).to contain_exactly(opened)
     expect(CashRegister.drawer_position_closed).to contain_exactly(closed)
+  end
+
+  it "builds inverse scopes with prefixed names for each namespaced element" do
+    opened = CashRegister.new.tap(&:open!)
+    closed = CashRegister.new.tap(&:close!)
+
+    expect(CashRegister.drawer_position_not_open).to contain_exactly(closed)
   end
 end